[refactor] HomePage UI 컴포넌트 분리 및 폴더 구조 정리
