+++ src/components/home/FavoriteSection.jsx
@@ -1,3 +1,5 @@
+import React from "react";
+
 function FavoriteSection() {
   return <div>FavoriteSection</div>;
 }
