+++ src/contexts/ThemeContext.jsx
@@ -1,12 +1,11 @@
 import React, { createContext, useContext, useState, useEffect } from "react";
+import PropTypes from "prop-types";
 
 const ThemeContext = createContext();
 
-// eslint-disable-next-line react/prop-types
 export function ThemeProvider({ children }) {
   const [isDarkMode, setIsDarkMode] = useState(false);
 
-  // 다크 모드 상태가 바뀔 때마다 바디에 적용함
   useEffect(() => {
     document.body.setAttribute("data-theme", isDarkMode ? "dark" : "light");
   }, [isDarkMode]);
@@ -17,5 +16,9 @@ export function ThemeProvider({ children }) {
     </ThemeContext.Provider>
   );
 }
+
+ThemeProvider.propTypes = {
+  children: PropTypes.node.isRequired,
+};
 
 export const useThemeContext = () => useContext(ThemeContext);
